+++ app/src/test/java/reikai/presentation/library/novels/NovelDynamicGroupingTest.kt
@@ -47,5 +47,6 @@ class NovelDynamicGroupingTest {
         downloadCount = 0,
         latestUpload = 0,
         chapterFetchedAt = 0,
+        lastRead = 0,
     )
 }